debug
debuug

+++ debug12/9.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #define PI 3.14159
 int main() {
-	float r;
-    float c, s;
-	scanf("%g", &r);
+	double r;
+    double c, s;
+	scanf("%lf", &r);
     c = 2 * PI * r;
     s = PI * r *r;
     printf("%.3f\n",c);
